DocStringExtensions

DocStringExtensions — Module.

Extensions for the Julia docsystem.

Introduction

This package provides a collection of useful extensions for Julia's built-in docsystem. These are features that are still regarded as "experimental" and not yet mature enough to be considered for inclusion in Base, or that have sufficiently niche use cases that including them with the default Julia installation is not seen as valuable enough at this time.

Currently DocStringExtensions.jl exports a collection of so-called "abbreviations", which can be used to add useful automatically generated information to docstrings. These include information such as:

simplified method signatures;

documentation for the individual fields of composite types;

import and export lists for modules;

and source-linked lists of methods applicable to a particular docstring.

Users are most welcome to suggest additional abbreviation ideas, or implement and submit them themselves. Julia's strong support for program introspection makes this a reasonably straight forward process.

Details of the currently available abbreviations can be viewed in their individual docstrings listed below in the "Exports" section.

Examples

In simple terms an abbreviation can be used by simply interpolating it into a suitable docstring. For example:

using DocStringExtensions

"""
A short summary of `func`...

$(SIGNATURES)

where `x` and `y` should both be positive.

# Details

Some details about `func`...
"""
func(x, y) = x + y

$(SIGNATURES) will be replaced in the above docstring with

# Signatures

```julia
func(x, y)
```

The resulting generated content can be viewed via Julia's ? mode or, if Documenter.jl is set up, the generated external documentation.

The advantage of using SIGNATURES (and other abbreviations) is that docstrings are less likely to become out-of-sync with the surrounding code. Note though that references to the argument names x and y that have been manually embedded within the docstring are, of course, not updated automatically.

Exports

EXPORTS

FIELDS

IMPORTS

METHODLIST

SIGNATURES

TYPEDEF

Imports

Base

Compat

Core

Note

This package is installable on Julia 0.4, but does not provide any features. Abbreviations can still be interpolated into docstrings but will expand to an empty string rather than the expected auto-generated content that it would on Julia 0.5 and above.

This allows the package to be added as a dependancy for packages that still support Julia 0.4 to allow them to begin using abbreviations without needing to conditionally import this pacakage.

source

Index

DocStringExtensions.EXPORTS

DocStringExtensions.FIELDS

DocStringExtensions.IMPORTS

DocStringExtensions.METHODLIST

DocStringExtensions.SIGNATURES

DocStringExtensions.TYPEDEF

DocStringExtensions.Abbreviation

DocStringExtensions.MethodList

DocStringExtensions.MethodSignatures

DocStringExtensions.ModuleExports

DocStringExtensions.ModuleImports

DocStringExtensions.TypeDefinition

DocStringExtensions.TypeFields

DocStringExtensions.alltypesigs

DocStringExtensions.arguments

DocStringExtensions.cleanpath

DocStringExtensions.comparemethods

DocStringExtensions.format

DocStringExtensions.getmethods

DocStringExtensions.getmethods!

DocStringExtensions.groupby

DocStringExtensions.groupby!

DocStringExtensions.isabstracttype

DocStringExtensions.isbitstype

DocStringExtensions.keywords

DocStringExtensions.methodgroups

DocStringExtensions.parsedocs

DocStringExtensions.printmethod

DocStringExtensions.url

Reference

DocStringExtensions.EXPORTS — Constant.

An Abbreviation to include all the exported names of a module is a sorted list of Documenter.jl-style @ref links.

Note

The names are sorted alphabetically and ignore leading @ characters so that macros are not sorted before other names.

Examples

The markdown text generated by the EXPORTS abbreviation looks similar to the following:

  - [`bar`](@ref)
  - [`@baz`](@ref)
  - [`foo`](@ref)

source

DocStringExtensions.FIELDS — Constant.

An Abbreviation to include the names of the fields of a type as well as any documentation that may be attached to the fields.

Examples

The generated markdown text should look similar to to following example where a type has three fields (x, y, and z) and the last two have documentation attached.

  - `x`

  - `y`

    Unlike the `x` field this field has been documented.

  - `z`

    Another documented field.

source

DocStringExtensions.IMPORTS — Constant.

An Abbreviation to include all the imported modules in a sorted list.

Examples

The markdown text generated by the IMPORTS abbreviation looks similar to the following:

  - `Foo`
  - `Bar`
  - `Baz`

source

DocStringExtensions.METHODLIST — Constant.

An Abbreviation for including a list of all the methods that match a documented Method, Function, or DataType within the current module.

Examples

The generated markdown text will look similar to the following example where a function f defines two different methods (one that takes a number, and the other a string):

```julia
f(num)
```

defined at [`<path>:<line>`](<github-url>).

```julia
f(str)
```

defined at [`<path>:<line>`](<github-url>).

source

DocStringExtensions.SIGNATURES — Constant.

An Abbreviation for including a simplified representation of all the method signatures that match the given docstring. See printmethod for details on the simplifications that are applied.

Examples

The generated markdown text will look similar to the following example where a function f defines method taking two positional arguments, x and y, and two keywords, a and the b.

```julia
f(x, y; a, b...)
```

source

DocStringExtensions.TYPEDEF — Constant.

An Abbreviation for including a summary of the signature of a type definition. Some of the following information may be included in the output:

whether the object is an abstract type or a bitstype;

mutability (either type or immutable is printed);

the unqualified name of the type;

any type parameters;

the supertype of the type if it is not Any.

Examples

The generated output for a type definition such as:

"""
$(TYPEDEF)
"""
immutable MyType{S, T <: Integer} <: AbstractArray
    # ...
end

will look similar to the following:

```julia
immutable MyType{S, T<:Integer} <: AbstractArray
```

Note

No information about the fields of the type is printed. Use the FIELDS abbreviation to include information about the fields of a type.

source

DocStringExtensions.alltypesigs — Method.

alltypesigs(sig)

Returns a SimpleVector of the Tuple types contained in sig.

source

DocStringExtensions.arguments — Method.

arguments(m)

Returns the list of arguments for a particular method m.

Examples

f(x; a = 1, b...) = x
args = arguments(first(methods(f)))

source

DocStringExtensions.cleanpath — Method.

cleanpath(path)

Remove the Pkg.dir part of a file path if it exists.

source

DocStringExtensions.comparemethods — Method.

comparemethods(a, b)

Compare methods a and b by file and line number.

source

DocStringExtensions.format — Method.

format(abbr, buf, doc)

Expand the Abbreviation abbr in the context of the DocStr doc and write the resulting markdown-formatted text to the IOBuffer buf.

source

DocStringExtensions.getmethods! — Method.

getmethods!(results, f, sig)

A helper method for getmethods that collects methods in results.

source

DocStringExtensions.getmethods — Method.

getmethods(f, sig)

Collect and return all methods of function f matching signature sig.

This is similar to methods(f, sig), but handles type signatures found in DocStr objects more consistently that methods.

source

DocStringExtensions.groupby! — Method.

groupby!(f, groups, data)

A helper method for groupby that uses a pre-allocated groups Dict.

source

DocStringExtensions.groupby — Method.

groupby(f, K, V, data)

Group data using function f where key type is specified by K and group type by V.

The function f takes a single argument, an element of data, and should return a 2-tuple of (computed_key, element). See the example below for details.

Examples

groupby(Int, Vector{Int}, collect(1:10)) do num
    mod(num, 3), num
end

source

DocStringExtensions.isabstracttype — Method.

isabstracttype(t)

Is the type t an abstract type?

source

DocStringExtensions.isbitstype — Method.

isbitstype(t)

Is the type t a bitstype?

source

DocStringExtensions.keywords — Method.

keywords(func, m)

Returns the list of keywords for a particular method m of a function func.

Examples

f(x; a = 1, b...) = x
kws = keywords(f, first(methods(f)))

source

DocStringExtensions.methodgroups — Method.

methodgroups(func, typesig, modname; exact)

Group all methods of function func with type signatures typesig in module modname. Keyword argument exact = true matches signatures "exactly" with == rather than <:.

Examples

groups = methodgroups(f, Union{Tuple{Any}, Tuple{Any, Integer}}, Main; exact = false)

source

DocStringExtensions.parsedocs — Method.

parsedocs(mod)

Parse all docstrings defined within a module mod.

source

DocStringExtensions.printmethod — Method.

printmethod(buffer, binding, func, method)

Print a simplified representation of a method signature to buffer. Some of these simplifications include:

no TypeVars;

no types;

no keyword default values;

? printed where #unused# arguments are found.

Examples

f(x; a = 1, b...) = x
sig = printmethod(Docs.Binding(Main, :f), f, first(methods(f)))

source

DocStringExtensions.url — Method.

url(m)

Get the URL (file and line number) where a method m is defined.

Note that this is based on the implementation of Base.url, but handles URLs correctly on TravisCI as well.

source

DocStringExtensions.Abbreviation — Type.

Abbreviation objects are used to automatically generate context-dependant markdown content within documentation strings. Objects of this type interpolated into docstrings will be expanded automatically before parsing the text to markdown.

source

DocStringExtensions.MethodList — Type.

The singleton type for METHODLIST abbreviations.

source

DocStringExtensions.MethodSignatures — Type.

The singleton type for SIGNATURES abbreviations.

source

DocStringExtensions.ModuleExports — Type.

The singleton type for EXPORTS abbreviations.

source

DocStringExtensions.ModuleImports — Type.

The singleton type for IMPORTS abbreviations.

source

DocStringExtensions.TypeDefinition — Type.

The singleton type for TYPEDEF abbreviations.

source

DocStringExtensions.TypeFields — Type.

The singleton type for FIELDS abbreviations.

source